+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - `csvtk`: 
       - near all commands skip empty file now. [#204](https://github.com/shenwei356/csvtk/issues/204)
       - the global flag `--infile-list` accepts stdin "-". [#210](https://github.com/shenwei356/csvtk/issues/210)
+  - `csvtk pretty`: rewrite to support wrapping cells. [#206](https://github.com/shenwei356/csvtk/issues/206) [#228](https://github.com/shenwei356/csvtk/issues/228)
   - `csvtk cut/fmtdate/freq/grep/rename/rename2/replace/round`: allow duplicated column names.
   - `csvtk csv2xlsx`: optionally stores numbers as float. [#217](https://github.com/shenwei356/csvtk/issues/217)
   - `csvtk xlsx2csv`: fix bug wheree `xlsx2csv` treats small number (padj < 1e-25) as 0. It's solved by updating the excelize package. [#261](https://github.com/shenwei356/csvtk/issues/201)